chor(tasklist): adjust IE9 polyfill loading
related to #CAM-5680
