+++ ui/tasklist/client/index.html
@@ -97,7 +97,9 @@
   <script type="text/javascript">
     // override the following line to include configuration
     var tasklistConf = {};
+  </script>
 
+  <script type="text/javascript">
     window.PLUGIN_DEPENDENCIES = $PLUGIN_DEPENDENCIES || [];
     window.PLUGIN_PACKAGES = $PLUGIN_PACKAGES || [];
